requirements.txt
Flask==3.1.2
google-generativeai==0.8.5
supabase==2.8.1
python-dotenv==1.1.1
cryptography==43.0.0
